Added clean architecture
