+++ Hahahub/app/src/main/AndroidManifest.xml
@@ -5,7 +5,7 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name="com.hahahub.ui.App"
+        android:name="com.hahahub.startup.App"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.Hahahub"
         tools:targetApi="31">
         <activity
-            android:name="com.hahahub.ui.MainActivity"
+            android:name="com.hahahub.startup.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
