+++ Hahahub/app/src/main/java/com/hahahub/data/db/JokeEntity.kt
