+++ Hahahub/app/src/main/java/com/hahahub/domain/constants/Constants.kt
@@ -1,0 +1,8 @@
+package com.hahahub.domain.constants
+
+object Constants {
+    const val CATEGORY_KEY = "category"
+    const val QUESTION_KEY = "question"
+    const val ANSWER_KEY = "answer"
+    const val SOURCE_KEY = "source"
+}
