+++ Hahahub/app/src/main/java/com/hahahub/startup/App.kt
@@ -1,0 +1,11 @@
+package com.hahahub.startup
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class App : Application() {
+    override fun onCreate() {
+        super.onCreate()
+    }
+}
