+++ Hahahub/app/src/main/java/com/hahahub/startup/MainActivity.kt
@@ -1,0 +1,22 @@
+package com.hahahub.startup
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.example.hahahub.R
+import com.hahahub.presentation.ui.joke_list.JokesListFragment
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+
+        if (savedInstanceState == null) {
+            supportFragmentManager.beginTransaction()
+                .replace(R.id.fragment_container, JokesListFragment())
+                .commit()
+        }
+    }
+}
