--- Hahahub/app/src/main/java/com/hahahub/ui/App.kt
@@ -1,11 +1,0 @@
-package com.hahahub.ui
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class App : Application() {
-    override fun onCreate() {
-        super.onCreate()
-    }
-}
