--- Hahahub/app/src/main/java/com/hahahub/ui/constants/Constants.kt
@@ -1,8 +1,0 @@
-package com.hahahub.ui.constants
-
-object Constants {
-    const val CATEGORY_KEY = "category"
-    const val QUESTION_KEY = "question"
-    const val ANSWER_KEY = "answer"
-    const val SOURCE_KEY = "source"
-}
